Make EnumAttributeAccessorImpl compatible with enum classes.

+++ Source/Urho3D/Scene/Serializable.h
@@ -165,7 +165,7 @@ public:
     {
         assert(ptr);
         const T* classPtr = static_cast<const T*>(ptr);
-        dest = (classPtr->*getFunction_)();
+        dest = (int)(classPtr->*getFunction_)();
     }
 
     /// Invoke setter function.
